backend/requirements.txt
asgiref==3.3.4
# asgiref==3.6
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
colorama==0.4.4
Django==3.2.4
django-autoslug==1.9.8
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
flake8==3.9.2
flake8-broken-line
flake8-isort
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
gunicorn==20.1.0
iniconfig==1.1.1
isort==5.8.0
mccabe==0.6.1
packaging==20.9
pdfkit==0.6.1
pep8-naming==0.11.1
Pillow
djoser
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest
python-dotenv==0.17.1
python-slugify==5.0.2
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
testfixtures==6.17.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
drf_base64==2.0
# django-cors-headers==4.4.0
